+++ packages/plugin-packages/mashroom-portal-sandbox-app/src/js/containers/PortalAppSelectionContainer.jsx
@@ -12,6 +12,7 @@ import type {
 import type {MashroomAvailablePortalApp} from '@mashroom/mashroom-portal/type-definitions';
 
 type OwnProps = {
+    preselectAppName: ?string,
     onSelectionChanged: (?string) => void,
 }